Bug 8714: move name of network interface to instance value in PCI tree
Instance names are not yet accessible via 'get' in configuration
files. Move the name of network interface of a PCI device into instance
value to fix that. Instance name for network interface
is either empty or indicates the sequence number of interface to easily
distinguish between 1 or many network interfaces per PCI device.