cleanup: add noreturn attribute to tapi_jmp_do()
In normal case the function never returns cause it does longjmp(). So if we
have if() else construct that does TEST_FAIL or default: with TEST_FAIL -
compiler should be able to tell that we're leaving the function there and say
avoid generating an error cause 'function has no return'.

Potential danger is that we actually have return statements there so compiler
might do false optimizations. Said that:

  o there is nothing we can do in those cases and we're not checking the
    return code,
  o and ERROR() is an IO operation that won't be optimized out by the
    compiler cause IO is an op with side effects.

So the change looks safe.