Bug 9846: add NUMA-aware CPU allocation TAPI
The CPU cores allocated for a DPDK application should be not only
isolated, but also be located on a NUMA node local to a PCI device.

Add TAPIs to make it easier to allocate CPU cores with a specific
topology, and a dedicated TAPI to allocate them on a single NUMA node.