Bug 11357: doc/cm/base: make agent/interface/mcast_link_addr volatile
Fix two problems:
1. Multicast address can appear during a test independently on a test agent
activity, what causes unexpected configuration changes and the test failure.
2. Configurator fails to roll back changes after tests execution on
one of configurations when an interface is moved to new network namespace.
In particular it fails (ENOENT) to delete a multicast address entry. It seems
the entry goes away when interface is moved to another network namespace and
back.