Bug 10902: fix live mode of RGT tool
If RGT is run in live mode before the current testing starts, it
can open RAW log of the previous testing and stop at its end. It
should instead detect that a new RAW log is available by comparing
inode of the currently opened file with inode of the file available
at RAW log location.
Also it should clear EOF flag before every read attempt to be able
to continue when new records are added to the current RAW log.