Bug 11360: engine/builder: make TA signatures user-configurable
By setting the TE_TA_SIGNATURE_FIELDS environment variable, the user can
control the information that is placed in TA signatures. This variable is a
string that contains a sequence of keys that represent different kinds of
information about a given host.

This patch defines the following keys:
1. "gcc-target": a GNU triplet as detected by GCC;
2. "libc-version": currently used version of libc;
3. "libc-kernel": kernel version that was used to build libc;
4. "header-version": Linux header version.

If TE_TA_SIGNATURE_FIELDS is not set by the user, these 4 keys will be used
in the given order.