Bug11373: missing instance "status" on agent interface tree
ethtool can return ENODEV for some interfaces if the last ones are not
active. For example, with r8169: Cannot get device settings: No such device

If to turn interface on, it is able to gather all needed info.

Hence, ENODEV it is not a case to terminate agent/interface
initialization