Bug 11397: meson: do not copy files from the list .gitignore using rsync
Pass .gitignore in rsync when copying TE sources from the engine to the host
(or temporary directory). This patch improves the changes made in the commit
aebd597420d3 but without using eval.
Note: the signature of the rsync_to() function is changed to allow additional
rsync arguments to be passed when it needed.