Bug 11411: lib/tad: fill in mismatch packet payload
If mismatch packets are requested by the test, mismatch data should
be reported in payload. In theory, mismatch layer still could be
filled in, but in theory data could be too short for the layer and
it requires careful handling. Don't care about it for now. It means
that only match layers are filled in and nicely reported in NDS.