Bug 11459: support autorestart of processes controlled by CS
Add support of autorestart to unix/conf/process. If the process has
a non-zero autorestart option set, a dedicated thread will periodically
(the period is equal to the value of the autorestart option) check if the
process is running and restart it if it does not.

This is useful if a process should run for a certain period of time
during testing, in case of crash it will be restarted. Also there will
be an ability to start a process with a specified periodicity.