Bug 11471: fix error codes in conf/process and conf/vm
Different error codes were used for situations when a process/vm
should not be reconfigured beacuse it is running. In some places
TE_RC was called with ETXTBSY, not TE_ETXTBSY which is a wrong usage.
This patch sets error codes to be TE_EBUSY instead of (TE_)ETXTBSY
everywhere for such situations.