Bug 11132: engine/logger: add a rule for artifact processing in live logs
This patch implements a new listener event called "artifact". These events
contain information about the artifact log message along with both test and
plan IDs of the test that produced the artifact.

In order to decide which test an artifact belongs to, the "test_progress"
rule keeps track of test execution times. This means that the test IDs will
be present in the event messages only if the "test_progress" rule is used
anywhere in the configuration file.