Bug 11132: engine/logger: allow an optional trailing slash in URLs
Trailing slashes are needed for compatibility with listeners that rely on
opinionated web frameworks like Django.

Specifically, Django wants all URLs to end with a slash, which in our case
leads to URLs like these:

https://listener.example.com/import/init/
https://listener.example.com/import/feed/?run=ID
https://listener.example.com/import/finish/?run=ID

While "beautiful" in some sense of the word, it is hardly semantically
accurate, and we shouldn't push Django's conventions on all listeners.

This patch introduces a user-set option that makes Logger insert a trailing
slash into all HTTP request URLs for a given listener.