Bug 11435: fix force parameter usage in is_reqs_expr_match()
Allow to disable a run item by !REQ_NAME only if that expression
is either the only one or a part of top-level AND expression.
In more complex cases it may work incorrect (for instance, X|!Y
with X=FALSE and Y=TRUE will disable a session with force=TRUE
when X may be TRUE for some item inside that session which should
be run).