Bug 11495: tapi_tad: fix IP calculation to work with odd segment len
IP calculation of an odd segment works only when it's last in a packet.
It's buggy for example in case of VXLAN packet carrying inner Ethernet
frame with small odd payload because it implies addition of a trailer
payload segment to satisfy minimum Ethernet frame size. Eventually it
results in second to last segment being with odd length.