Bug 8744: add support for detecting TSO cutoff barrier
A HW TSO implementation may miss a means to gracefully
handle the case when its provisional buffer size isn't
a multiple of MSS while being smaller than the payload
size in the original superframe. The data chunk at the
end of the buffer gets transmitted as a separate frame.

Make the "Match Rx burst on TST with DPDK" tool detect
the TSO cutoff barrier effect and handle it gracefully.