Bug 11528: tapi_performance: fix iperf report function in polling case
The iperf get report function may be used with polling, on
entry it reads tool output into application buffer if it's
empty, and then tries to find some pattern in it, failing if
the pattern not found without free the buffer. The subsequent
calls to the function fail if the first call failed since the
same buffer content is used. Fix the function by appending
stdout messages to the buffer.