Bug 11550: engine/configurator: support backup verify/restore for direct DH
Typically, dynamic history is dropped after backup verify/restore since
we just checked that the current state is equal to the state in the
backup. However, when backup verify/restore is used on forward DH processing to
restore TA configuration after death and reboot, the verification is
done for the agent subtree only and we're going to apply subsequent dynamic
history commands to achieve the final target state. So, these commands
must not be dropped.