Bug 11550: Add the TA restore by direct DH commands
Restoring an agent by backup is not always possible, since a backup is a
snapshot of the final state. Dynamic history allows to approach this state
consistently. The approach is not ideal as well, since it assumes the
same initial state. Sometimes it is true (e.g. if testing runs just
after reboot), but sometimes it is false and in this case remaining bits
should be restored via backup.