Bug 11533: Store value with substitutions in DH
After a new instance was added or a new value for the instance was set,
the actual value of the instance could change after synchronization.
This could have happened because of the substitution mechanism.
Therefore, commands with the actual values of instances should be added
to the dynamic history after synchronization.