Bug 11459: introduce TA job library for agent job control
The library contains code from lib/rpcs_job/job.c (it will be removed
from there with one of the next patches) with minor changes that are required to
make it a suitable library code.
The library is intended to be used by TAPI job backends like the RPC
backend (lib/rpcs_job/job.c) and a Configurator backend which will be
based on agent/unix/conf/process/conf_process.c (will be added in future
patches).
The patch does not change any current behaviour.