Bug 11459: TA job: do not free argv and env if job creation was unsuccessful
ta_job_create() will not free argv and env if it is unsuccessful.
Callers of the function must free them themselves.
This is done because otherwise it is not clear whether a caller
of ta_job_create() should free the arguments.
Also use te_str_free_array() for freeing because it does the job well.