Bug 11595: add --txt-mi-raw rgt/log.sh to include raw MI into TXT log
This is useful when you want to see RAW json instead of some post-processing
that was done by TE. Which in turn can be used to take MI artifacts and
post-process those.

Proper approach would be to add RGT filter that will allow to include only MI
artifacts, but short term a hack is good enough + it has it's own value.