Bug 11609: lib/tools: bypass log level check when console is logged
Console messages log level is specified in various configuration files
(either RCF, or Logger, or CS for log messages parser case).

WARN level is too verbose and typically highlighted very aggressively
in HTML logs. Logically messages in serial console are just an extra
information. It should not be treated as warnings. However, INFO level
is disabled by default. So, bypass check when console messages are
logged in assumption that user really wants to see it if console
logging is enabled.