Bug 11609: support log user control for serial console parser
Console configuration is a bad log user since it could be really long.
Yes, it matches conserver console name in some cases, but sometimes it
is a netconsole configuration with random ports, sometimes it is not
unique (e.g. if it is just 'netconsole:' when port is configured
separately), sometimes it is a serial device name.