Bug 11619: purge unused libc-kernel variant for get_ta_signature
This variant was used by sapi-ts, but it stopped from it this September.

In any case, this libc-kernel meant "libc was built with headers taken
from linux kernel of this version".  Glibc uses its own copy of linux
headers for many years now, so this libc-kernel version has no sense.