Bug 11606: engine/rcf: do not reboot agents upon check
RCF can receive a request to check agents (e.g. the configurator).
If RCF detects that the agent is dead, it tried to reanimate it
automatically and then the configurator tried to restore
configuration backup. But this does not guarantee a successful
recovery. Since the configurator controls restore process, it
should be responsible for the reboot control as well.