Bug 11637: switch default behaviour with SIGUSR2 signal in test
Previously SIGUSR2 would stop all the testing completely, and there was
--test-sigusr2-verdict to print a verdict in the current test. This patch
switches the default. Now the default behaviour is to print a verdict in
the test, and an option --test-sigusr2-stop can be used to stop all the
testing completely.
Also print the verdict "Test caught the SIGUSR2 signal" instead of "Test
is stopped by SIGUSR2", since the test is not actually stopped.