Bug 11533: lib/rcfrpc: add processing ENOENT error
/agent/rpcserver/sid is the volatile object and can be deleted after
synchronization with TA in the get CS handler. This can happen if
something went wrong in the previous iteration of the test and RPC
server is dead. For this reason, getting the sid value may fail with
an ENOENT error. But this is not a fatal error, we should just
detect it and continue restarting the RPC server if the reuse_pco is set.