Bug 11568: use systemctl for ntp server
On some new distributions (Debian 11) \'service\' tool calls
systemctl which prints logs at the end of the status output.
We are not interested in the logs, and reading logs takes up to
a minute after fresh reboot.

This patch adds direct call to systemctl when handling ntp service,
and passes `-n0` parameter so that logs are not printed.