Bug 10129: stop reloading modules from TA dir
There is no need to reload a module if it was already taken from TA's
directory. But it's especially important when we want our IUT to be a VF:
in that case we have to set up VFs before the tests are run, and
unloading the module behind the backing PF will destroy that setup.