Bug 11088: Add the ability to set specified l4 port
After the port has been allocated, it can be used by another agent on
another host to setup data exchange(for example, netconsole). After the reboot,
it is expected that the configurator will be able to restore the agent.
The agent must use the same ports as before the reboot to try to restore
connections. Therefore, the l4 port tree should be able to allocate the
specified port.