doc: TAPI RPC: emphasize that there's no need to copy input arrays
It is worth mentioning that the arrays which are input arguments for
rcf_rpc_call() do not have to be copied.
The patch also adds a reference to the section where this is discussed
to the beginning of the page for users to find it quicker and not to
miss it.