Bug 11459: add type-safe versions of cfg_get_instance_* functions
This way a warning will occur if the passed value pointer is of a wrong
type, which is beneficial for type safety.
Note that the functions do not check if the type of the received value
is the one we asked for because the check is done in cfg_get_instance().