Bug 8744: tapi_dpdk: fix set mbuf size when adjust defaults
Mbuf size must be based on maximum of maximum segment size and
packet size. For Tx it could be correct to use maximum segment
size. But since RxQ is set for Tx side too, mbuf size must be
set with respect to it, i.e. we should consider packet size.

Fixes: 48775c9b6a83 ("Bug 8744: tapi_dpdk: set default testpmd parameters")