Bug 11672: do not complain when device parameters cannot be obtained
Currently pci_param_list() can return TE_ENOENT if it cannot list
devlink parameters (for example, because the required devlink command
is not supported at all). This results in printing a lot of messages
telling that "List method failed" in log. It would be better to return
empty list to avoid these messages.