Bug 11459: fix a function that removes filters to avoid use-after-free
tapi_job_filter_remove_channels() should check if the filter is attached
to the channels which are being removed.
Otherwise, the following scenario was possible: the filter is attached
to stdout, tapi_job_filter_remove_channels() is called to remove the
filter from stderr. This leads to the filter being freed, so the further
attempts to use it cause undefined behaviour.