Bug 11766: save workspaces for TCE processing
TA agents run on distinct hosts, TA hosts. Some of this TA hosts,
platform hosts, are used to build TA agents. To build a TA agent on a
platform host a dedicated directory, workspace, is used.

It is required to use this TA workspaces to process TCE information:
- On a TA agent, TCE runtime information, .gcda files, are created in
  the directory with the same path as the path to the workspace
  directory of the platform host used to build this TA agent.
- Each platform host provides appropriate environment, including the
  workspace directory, to process .gcda files received from TA agents
  build on this platform host.

Paths to workspaces, local to platform hosts, are kept in files on the
TE host. There are some environment variables to locate this files,
but providing this paths as a Bash array simplifies them usage.

Thus, workspaces is read only once and can be reused by further
commits.