Bug 11766: store build paths to process TCE from TA agents
To build a TA agent on a platform host a dedicated directory, workspace,
is used. TA agent consists from a set of components. Each component is
built in a distinct subdirectory of the workspace directory.

Let a component provides TCE information. This information provides
paths relative to the build directory. And the path of this directory
should be used to process this information:
- The gcov utility should be run from this directory.
- The tce_summary script should be run from this directory.
- The path to the build directory relative to the workspace directory
  is kept the same between distinct platform hosts and can be used as
  the reference point to process absolute paths.

There are some environment variables to locate build paths, but
providing this paths as a Bash array simplifies them usage.

Further commits will use this paths to process TCE for each host.