Bug 11766: fetch TCE from TA agents
TA agents are built on some hosts, platform hosts, and run on others, TA
hosts. Indeed, TCE information, gcov data files, is created TA hosts.

To process TCE information by the gcov utility the following files
should be provided at the same paths as those used to build this TA
agent on the correspondent platform host:
- gcov graph files,
- source files,
- header files if contain 'static inline' functions.

All this files are already present on the platform host used to
build a TA agent. So that, it is reasonable to process TCE information
on the platform host.

Thus, further commits will process fetched TCE information.