Bug 11766: clean TCE on platform
TA agents run on distinct hosts, TA hosts. Some of this TA hosts,
platform hosts, are used to build TA agents.

To process TCE information by the gcov utility the following files
should be provided at the same paths as those used to build this TA
agent on the correspondent platform host:
- gcov graph files,
- source files,
- header files if contain 'static inline' functions.

All this files are already present on the platform host used to
build a TA agent. So that, it is reasonable to process TCE information
on the platform host.

Let some TA agents is built on the same platform host. In general, each
of this TA agents provides a distinct set of gcov data files. So that,
to remove all gcov data files on this platform hosts is the simplest way
to grant that only gcov data files from a single TA agent will be
processed at once.

Thus, further commits will use this platform to process TCE already
fetched from TA agents.