Bug 11459: change return type of rpc_job_* functions from int to te_errno
The change is required for unification: some tapi_job_* functions are
going to call both rpc_job_* and tapi_cfg_ps_* functions, so it would be
convenient if they had the same signature.

The return values we get from RPC server are te_errno-s packed into
tarpc_int which are implicitly converted to int via RETVAL_INT().
The values are used in tapi_job_* functions as te_errno-s, so we already
rely on the fact that te_errno and int are the same.