Bug 11459: change job-related rpc_job_* functions not to be RPC specific
rpc_job_* functions that work with particular job are going to be used
as methods. This means that tapi_job_t structure will contain pointers
to the functions so that instead of e.g. rpc_job_create() we will
use job->create(). This will allow us to use different backends, not
just RPC. But this requires the backend functions to have signatures
that are not specific to the backend.