Bug 9206: Make the loaded node of the module tree non volatile
Volatile nodes are not added to backup and DH, so they cannot be
restored. In order for a module to be unloaded, loaded node must be non
volatile. However, because of the node was volatile, and two agents
could use the same module at the same time without any problem, since
the module was not unloaded. For this reason, this patch adds the
ability to make a fake unload if the module is occupied by another
agent.