Bug 11766: create TCE summary
TA agents are built on distinct hosts, platform hosts. To build a TA
agent on a platform host a dedicated directory, build, is used. TCE
information, TE TCE files, is relative to the build directory of a
platform chosen to merge this information.

To provide TCE summary the tce_summary utility is used now. This utility
must be called from the directory, build directory, paths within TE TCE
files are related to.

Thus, cross-platform TCE summaries are provided for each TCE component.