Bug 11972: implement sequence numbers for RPC calls
In some cases it may be necessary to correlate logs
regarding RPC calls from TAPI and the agent, e.g when
some state related to the RPC call is only available at
the agent side. Since timestamp ordering is imperfect
and there may be additional log message interspersed
between a TAPI-side and an agent-side logs, some form
of a RPC call identifier is needed. The patch provides
such identifiers.