Bug 8552: tarpc_poll: don't log pollfd before call
The only parameter that is changed with poll() function(man poll) is
revents. And it's changed after the call. Developer can't make changes
to it before call, because it's an output parameter.
Patch suggests to log pollfd after the call and in case any
events are put into revents parameter - add prefix after the events.