Task 11973: cm_base: add module and binary directories to TA configuration model
The 'lib_mod_dir' is intended to store a directory for kernel module files
inside TA directory, while  'lib_bin_dir' stores a directory with  library
files (also in the TA directory).
It is assumed that the script that builds modules and libraries as part of
TE will put them in the same directories.
Also change TAPI to load self-built kernel modules from TA module directory.