Bug 12002: allow to copy VCS-related files to the sources directory for build
Some tools, for example Onload, require .git directory to build. The new
option --with-vcs-stuff can be used in TE_PLATFORM_EXT and TE_TA_APP macros.
If this option is not specified, then the directory .git and file .gitignore
will not be copied (as it was before this patch).
Note: this patch breaks the build on engine hosts with bash version lower
than 4.3 (due to the use of nameref variables).