Bug 12002: builder: delete excluded files from destination dirs
The builder can use the rsync --exclude argument, so we need to be sure that
there will be no unnecessary things in the target directory.  This may occur
after removing the --with-vcs-stuff directive in the TE_PLATFORM_EXT or
TE_TA_APP macro, because without this patch, the .git directory and .gitignore
file will remain.