Bug 11811: agents/unix/conf/base: make current channel counts volatile
Some drivers may override and protect these counts during
runtime. In example, mlx5 is known to do that when AF_XDP
zerocopy sockets are registered. Configurator discovers
this and tries to restore previously known values. It
fails to do so, presumably, because of the protection.

As a workaround, make the current channel counts volatile.