Bug 12068: extend and improve ta_cfg_obj implementation
Make it possible to use ta_cfg_obj to cache objects not only for
commit() methods but also for a group of get requests. Configurator
issues such a group when it synchronizes a tree. When all the properties
of an object are retrieved with a single system call, it is more
efficient to cache whole object when the first get request is processed
and then retrieve values from that cache until request group ID
increments.