lib/rpcserver: return a more accurate error on too many iovs
According to POSIX spec readv/writev may fail as follows:
    EINVAL The iovcnt argument was less than or equal to 0,
            or greater than IOV_MAX
but ENOMEM was returned by RPC wrappers which is inconsistent and
clearly incorrect. It is very unlikely the change is going to break
anything.

The present patch is an amalgamation of two patches that were pushed
and later reverted. No other changes have been made to them after
they were approved on Reviewboard.

OL-Redmine-Id: 12395
Signed-off-by: Nikita Grebennikov <nikita.grebennikov@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>