lib/conf_oid: add a function to dispatch on an object OID
In many cases it is necessary to perform different actions
depending on the object OID of a given node. In the existing
code it is usually done by a cascade of ifs and string
comparisons which is neither very readable nor robust.
This patch provide a clean uniform function to perform such tasks.

OL-Redmine-Id: 12442
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>