cleanup: remove autotools build support for TA platforms
It is very hard to support two build systems in parallel.
Supported meson versoin is available or easy to get on all
support distros.

Keep autoconf/automake files in not yet migrated agent types and
tools/milicom just for the reference to simplify migration in the
future.

Autotools-based build of test suites is still supported.