cleanup: remove file authors
Having authors in a file headers provokes corresponding changes when
a new contributor changes the file. The boundary when a new author
should be added is very hard to define. So, it is better to avoid
the practice. Therefore, it is decided to remove these lines before
making TE public.