lib/tapi_tad: remove initial list from IGMPv3 macro params
The optional initial values in the IGMPv3 list creating macros are
redundant and not quite safe. For example - the first parameter of the
IGMP3_GROUP_RECORD macro provokes to create a Group Record on stack.
If such entries are added to the Group Records List,
memory corruption will occur when the list is freed.

Signed-off-by: Anatoly Mirin <anatoly.mirin@oktetlabs.ru>
Reviewed-by: Ivan Melnikov <ivan.melnikov@oktetlabs.ru>
Reviewed-by: Ivan Malov <ivan.malov@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>