lib/rcfpch: do not log error on file open failure
The message in question may look confusing in logs when
one tries to append data to a non-existent file which
gets created later (see tapi_file_append_ta() and its
use in tapi_cfg_key_append_public() for an example).

When a file cannot be retrieved from TA, it
is TAPI wrappers who are supposed to log
the error, not rcf_pch_file() itself.

OL-Redmine-Id: 11932
Signed-off-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>