cs: make integer data type to be of 32-bit fixed width
Keeping in mind that CPU arch could differ on TE and TEN, it is rather
strange to have 'int' as the data type for integers in CS. Which size
is it?

We switched int to int32_t and then (in following commits) we will
define other integer types like uint8_t, _Bool and so on.

Ideally at least all get API usages which provide 'int' storage to get
value should be updated to use 'int32_t'. However it is not the subject
of this patch. It should be fixed in follow up patches.

Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@oktetlabs.ru>