engine/configurator: use dedicated field to link backup list
Currently, Configurator uses the same field (brother) for linking
instances in a backup list as normally used for linking list of
children of some instance. This makes it impossible to store
information about parent-child relations in the backup list,
but such will be needed for handling objects with unit=true.

OL-Redmine-Id: 9403
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>