lib/bpf: fix BPF map definitions
Legacy map definitions using bpf_map_def structure are going to be
deprecated and eventually dropped. However new (BTF-defined) definitions
are not fully supported on older systems. For example, on Debian 10
(clang 7.0.1-8) they can be compiled but libbpf refuses to load
compiled programs.

To resolve this issue, add detection of new map definitions support
in te_build_bpf and a macro defining BPF map in the legacy way if
the new way is not supported.

Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>