lib/confapi: make uniform names of conversion functions
The names like addr2str() or int2str() is very convinient and clear.
However when uint64 was added the name was chosen uint64_to_str().
I've switched the names to addr_to_str() and so on for uniformity and
to be able to use macroses for these function in the next commits.
Also functions for strings also have been made to have uniform names.
'char' or 'str' switched to uniform 'string'.

OL-Redmine-Id: 11890
Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@oktetlabs.ru>