build: bump Meson version
We are not going to support Ubuntu 18.04 anymore, so the earliest
version of Meson that can be met in the wild is 0.49.0 (found in
Debian stretch-backports). On the other hand, recent commits require
features of Meson >= 0.47.1 causing build warnings.

This patch mandates the use of Meson >= 0.49.0 and thus eliminates
warnings.

Some scripts have different special handling of Meson < 0.46 ---
all this code is now redundant and deleted.

'build_always' option was deprecated in Meson 0.47, so it is replaced
by a recommended alternative to avoid warnings.

OL-Redmine-ID: 12330
Fixes: 7deafd1b3b04 ("build: add missing package info display to meson")
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>