engine/builder: generate TCE configuration for RCF controller
The TCE information is generated on TA agents within the same directories
the TA agents have been built. The configuration provides paths to this
directories and to the directory, the TCE workspace directory, to save the
TCE information into.

The RCF controller will use this configuration to save the TCE information
onto the TE host. Also, the TCE workspace directory is provided before the
RCF controller is started.

OL-Redmine-Id: 11766
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Acked-by: Artem Andreev <artem.andreev@oktetlabs.ru>