lib/tapi_tad: allow to create an empty IGMPv3 report
This is necessary to create a pattern to capture reports with any group
records and later check if all the required reports have arrived.

IGMPv3 reports can contain multiple group records, with or without
auxiliary data. When a report is broadcast to an upstream network,
it can be somehow modified - divided into several reports,
discarding auxiliary data, etc.

Signed-off-by: Anatoly Mirin <anatoly.mirin@oktetlabs.ru>
Reviewed-by: Ivan Melnikov <ivan.melnikov@oktetlabs.ru>
Reviewed-by: Ivan Malov <ivan.malov@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>