common: follow C99 in int-to-pointer conversion
The only valid (as per ISO C99) way to convert a pointer to an integer
is by using intptr_t or uintptr_t. Those are the only two types that are
guaranteed to hold all possible pointer values. In particular, casting
a integer value to long and then to a pointer, in order to eliminate
compiler warning about casting from an integer of different size to
a pointer (like: int x; void *p = (void *)(long)x) is unreliable,
because long is not required to be the same size as a pointer, in theory
it may be shorter or it may be longer.

OL-Redmine-Id: 12469
Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>