lib/rcfunix: request to save TCE on reboot and shutdown
The following cases should be considered:
- the RCF controller is requested to shutdown a TA agent,
- the RCF controller is requested to reboot a TA agent.

In the former case the operation 'finish' is called on the RCF/
controller connected with the TA agent that should be shutdown.
This operation:
- terminates the TA agent,
- destroyes the RCF/ controller.
So that, the TCE information can be fetched on the 'finish' operation.

In the later case the behaviour depends on the configuration of a
specific TA agent:
- If the 'soft' reboot is allowed the 'finish' operation is called.
- If the 'host' reboot is allowed, the 'reboot' command is send to
  the TA agent.
So that, to process the case when the 'host' reboot is allowed only,
the 'reboot' command should be caught to save the TCE information.

Further commits will add text to fetch gcda files from a TA agent when
it is requested.

OL-Redmine-Id: 11766
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>