suites/selftest: use simpler checks for I/O ops success
Under normal circumstances, a blocking I/O on a regular file shall
not read or write less data than requests (unless the file is just
not long enough). So the removed macros are misguided and needlessly
overcomplicated, it is sufficient just to check that the returned
length is the same as expected, and fail the test if it is not.

OL-Redmine-Id: 12305
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>