lib/tapi: relocate filename-generating functions
There's not a single reason why these filename-generating functions
should be inline.

Note, however, that the patch may break things for users who may
have erroneously called tapi_file_generate_name() and then some other
File TAPI function and expected the pointer to a name remain valid.
This has never been guaranteed, so it's always a bug on the side
of the caller.

OL-Redmine-Id: 12117
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Ivan Melnikov <ivan.melnikov@oktetlabs.ru>