tools/tce: update DATA field of TCE reports on merging
To provide TCE summaries for each TA agent and for each TA type the TCE
information is represented as the TCE reports and then merged.

One gcda file provides the TCE information for the source file and
possibly many files included within it. The TCE report keeps both gcda
and source paths.
To merge the TCE information a `tce_merge` utility is used. This utility
requires gcda and source paths of files merged to be equal.

Thus, make the `tce_merge` utility to be happy.

OL-Redmine-Id: 11766
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Acked-by: Artem Andreev <artem.andreev@oktetlabs.ru>