tools/tce: update TCE summary utility does not lose TCE
A TCE summary represents a TCE report as a single HTML document. To
generate the TCE summary an utility, `tce_summary`, is used.
This utility provides a few modes to order the TCE information within the
TCE summary.

Thus, does not lose TCE information if some files has the same values of
the parameter used to order the TCE information.

OL-Redmine-Id: 11766
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Acked-by: Artem Andreev <artem.andreev@oktetlabs.ru>