tools/trc: support new format for specifying key URLs
Currently a file specified in --key2html option is a collection
of regular expressions with '\t' used as a delimeter between
expression and replacement string. It is not convenient to edit it.

Simpler format is introduced which starts with "#KEY_DOMAINS" comment.
Every line after that should contain key domain and URL prefix
(separated by spaces). It will tell TRC tools to replace
"ref://[domain]/" with specified URL prefix in key references.

OL-Redmine-Id: 12418
Signed-off-by: Dmitry Izbitsky <dmitry.izbitsky@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>