lib/tools: add definitions for PCI-related constants
Subsequent patches will make use of PCI class and subclass
values, so it is better to define them in a single place
in a concise manner, rather than using ad hoc definitions or
magic numbers.

So the definitions for PCI classes, subclasses and programming
interfaces are extracted from the public PCI ids database and
converted to enum values.

OL-Redmine-Id: 12444
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>