lib/tools: add function to extract varying part of filename
The function is intended to complement te_file_scandir()
from the previous patch. It allows a user to extract the
varying part of a filename matching a given glob pattern.
So the same pattern may be used for directory scanning and
extracting. The pattern must contain strictly a single wildcard.

OL-Redmine: 12447
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>