engine/builder: refactor dying operation for TCE processing
The TA hosts can be rebooted or destroyed. Thus, the fetching of the TCE
information should be controlled by the RCF controller. To achieve this
the controller should be provided with some additional information.

Further commits will reuse text dependant on this operation to provide
the RCF controller with information about TCE components.

OL-Redmine-Id: 11766
Signed-off-by: Boleslav Stankevich <boleslav.stankevich@oktetlabs.ru>
Acked-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Acked-by: Artem Andreev <artem.andreev@oktetlabs.ru>