lib/tools: add functions to stop and restart a timer
te_timer_start() does not allow to update a timer's timeout
and it is correct to avoid accidental timer rearm which
may be subtle and difficult to detect.

On the other hand we need a stuff which allows to restart
timer before it expires, for intsance in some repeated routines.

OL-Redmine-Id: 11566
Signed-off-by: Ivan Melnikov <ivan.melnikov@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>